Reduce HashMap entry sizes from size_t to uint for better cache density
Degree counts and simplex counts won't exceed 2^32. Shrinking entries
from 8 to 4 bytes improves cache locality in doMove (34 μs → from 38 μs,
~11% improvement). Also makes modifyFVector generic over element type
and adds move-type and delta-GC stats to benchmarks.

+++ source/applications/manifold_sampler.d
@@ -375,7 +375,7 @@ real speculativeBistellarDelta(int dim, Vertex, P)(
     allVerts.sort();
 
     // Compute speculative f-vector
-    size_t[dim + 1] newFVector = mfd.fVector[0 .. dim + 1];
+    uint[dim + 1] newFVector = mfd.fVector[0 .. dim + 1];
     newFVector[].modifyFVector(move);
 
     // Compute speculative totSqDeg for dimensions 0 through dim-2.
